scraper/requirements.txt
requests
beautifulsoup4
pdfminer.six
playwright
pydantic
python-dateutil
pytz
